WIP Fetch refactor
